+++ app/android/app/src/main/kotlin/com/flowdo/app/MainActivity.kt
@@ -1,5 +1,0 @@
-package com.flowdo.app
-
-import io.flutter.embedding.android.FlutterActivity
-
-class MainActivity : FlutterActivity()